refactor: move saniztize probe function
