+++ src/looper/index.ts
@@ -26,7 +26,6 @@ import { doProbe } from '../components/probe'
 import { getContext } from '../context'
 import type { Notification } from '@hyperjumptech/monika-notification'
 import type { Probe } from '../interfaces/probe'
-import { log } from '../utils/pino'
 import {
   getProbeContext,
   getProbeState,
@@ -38,95 +37,6 @@ export const DEFAULT_THRESHOLD = 5
 let checkSTUNinterval: NodeJS.Timeout
 
 const DISABLE_STUN = -1 // -1 is disable stun checking
-
-export function sanitizeProbe(isSymonMode: boolean, probe: Probe): Probe {
-  const { id, name, requests, incidentThreshold, recoveryThreshold, alerts } =
-    probe
-  probe.alerts = alerts?.map((alert) => {
-    if (alert.query) {
-      return { ...alert, assertion: alert.query }
-    }
-
-    return alert
-  })
-
-  probe.requests = requests?.map((request) => {
-    if (!request.method) {
-      return { ...request, method: 'GET' }
-    }
-
-    return {
-      ...request,
-      alerts: request.alerts?.map((alert) => {
-        if (alert.query) {
-          return { ...alert, assertion: alert.query }
-        }
-
-        return alert
-      }),
-    }
-  })
-
-  if (!name) {
-    probe.name = `monika_${id}`
-    log.warn(
-      `Warning: Probe ${id} has no name defined. Using the default name started by monika`
-    )
-  }
-
-  if (!incidentThreshold) {
-    probe.incidentThreshold = DEFAULT_THRESHOLD
-    log.warn(
-      `Warning: Probe ${id} has no incidentThreshold configuration defined. Using the default threshold: 5`
-    )
-  }
-
-  if (!recoveryThreshold) {
-    probe.recoveryThreshold = DEFAULT_THRESHOLD
-    log.warn(
-      `Warning: Probe ${id} has no recoveryThreshold configuration defined. Using the default threshold: 5`
-    )
-  }
-
-  if (alerts === undefined || alerts.length === 0) {
-    const getDefaultAlerts = (isHTTPProbe: boolean) => {
-      if (!isHTTPProbe) {
-        return [
-          {
-            assertion: 'response.status < 200 or response.status > 299',
-            message: 'Probe is not accesible',
-          },
-        ]
-      }
-
-      return [
-        {
-          assertion: 'response.status < 200 or response.status > 299',
-          message: 'HTTP Status is {{ response.status }}, expecting 200',
-        },
-        {
-          assertion: 'response.time > 2000',
-          message:
-            'Response time is {{ response.time }}ms, expecting less than 2000ms',
-        },
-      ]
-    }
-
-    const isHTTPProbe = requests.length > 0
-    probe.alerts = getDefaultAlerts(isHTTPProbe)
-    if (isHTTPProbe) {
-      log.warn(
-        `Warning: Probe ${id} has no Alerts configuration defined. Using the default response.status != 200 and response.time > 20000`
-      )
-    }
-  }
-
-  if (isSymonMode) {
-    probe.alerts = []
-  }
-
-  return probe
-}
 
 export async function loopCheckSTUNServer(interval: number): Promise<any> {
   // if stun is disabled, no need to create interval
